refactor: separate docker stuff
